BUG: make L-BFGS-B work with single precision gradient

+++ scipy/optimize/_lbfgsb_py.py
@@ -183,6 +183,9 @@ def fmin_l_bfgs_b(func, x0, fprime=None, args=(),
         fun = func
         jac = fprime
 
+    # Fortran expects a double precision array for gradient
+    jac=jac.astype(np.float64)
+
     # build options
     callback = _wrap_callback(callback)
     opts = {'disp': disp,
@@ -196,8 +199,8 @@ def fmin_l_bfgs_b(func, x0, fprime=None, args=(),
             'callback': callback,
             'maxls': maxls}
 
-    res = _minimize_lbfgsb(fun, x0, args=args, jac=jac, bounds=bounds,
-                           **opts)
+    res = _minimize_lbfgsb(fun, x0, args=args, jac=jac.astype(np.float64),
+                           bounds=bounds, **opts)
     d = {'grad': res['jac'],
          'task': res['message'],
          'funcalls': res['nfev'],
@@ -348,6 +351,8 @@ def _minimize_lbfgsb(fun, x0, args=(), jac=None, bounds=None,
     n_iterations = 0
 
     while 1:
+        # convert gradient to double precision for Fortran
+        g = g.astype(np.float64)
         # x, f, g, wa, iwa, task, csave, lsave, isave, dsave = \
         _lbfgsb.setulb(m, x, low_bnd, upper_bnd, nbd, f, g, factr,
                        pgtol, wa, iwa, task, iprint, csave, lsave,
